CRUD za foldere

+++ Mail/src/main/java/uns/ac/rs/Mail/repository/FolderRepository.java
@@ -6,6 +6,6 @@ import uns.ac.rs.Mail.entity.Folder;
 
 import java.util.List;
 
-public interface FolderRepository extends JpaRepository<Folder, Integer> {
+public interface FolderRepository extends JpaRepository<Folder, Long> {
 //    List<Folder> findByAccount(Account account);
 }
